refactor: extract Repo.Filtering module and shared changeset_for helper
Extract ~150 lines of filtering logic (query building, filter parsing,
pagination, ordering, soft-delete) from Ectomancer.Repo into a focused
Ectomancer.Repo.Filtering module with delegation stubs for backward compat.

Also introduce a shared changeset_for/4 helper that replaces 5 duplicated
changeset-building patterns (checking function_exported? for changeset/2
and falling back to Ecto.Changeset.cast).
